Bugfix in vcat.
(cherry picked from commit 51b6d24dfbc64d746cf592595b6532eaa5ca68aa)
ref #13315

+++ base/abstractarray.jl
@@ -738,7 +738,7 @@ function hcat(X::Number...)
 end
 
 function vcat{T}(V::AbstractVector{T}...)
-    n = 0
+    n::Int = 0
     for Vk in V
         n += length(Vk)
     end
